Move get_accumulated_output_amounts, comments
